feat: load halloween in require, close #15
