+++ cypress.json
@@ -1,6 +1,5 @@
 {
   "projectId": "ipnbuo",
   "viewportWidth": 1200,
-  "viewportHeight": 1200,
-  "theme": "halloween"
+  "viewportHeight": 1200
 }
